organization: Moved classes regarding login feature to login package #244
