+++ application/backend/src/main/java/cmpe451/group12/beabee/config/security/WebSecurityConfig.java
@@ -1,6 +1,6 @@
 package cmpe451.group12.beabee.config.security;
 
-import cmpe451.group12.beabee.util.JwtRequestFilter;
+import cmpe451.group12.beabee.login.util.JwtRequestFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -18,8 +18,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import javax.persistence.Id;
 
 @Configuration
 @EnableWebSecurity
