+++ application/backend/src/main/java/cmpe451/group12/beabee/login/model/Users.java
@@ -1,16 +1,12 @@
-package cmpe451.group12.beabee.model;
+package cmpe451.group12.beabee.login.model;
 
 import lombok.*;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import javax.persistence.*;
 
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.List;
-import java.util.Set;
 
 @Entity
 @Getter
