+++ application/backend/src/main/java/cmpe451/group12/beabee/login/service/LoginService.java
@@ -1,0 +1,2 @@
+package cmpe451.group12.beabee.login.service;public class LoginService {
+}
